refactor: move tmp_dir out of the git module
`tmp_dir` has nothing to do with git — config, lib and the clone path
all call it, and it stayed compiled without the `git` feature anyway.
It lives in `src/utils.rs` now, so the git facade only exports things
that are actually about git.
